+++ src/encryption.rs
@@ -3,7 +3,6 @@ use std::time::Duration;
 use matrix_sdk_common::deserialized_responses::{
     ShieldState as RustShieldState, ShieldStateCode as RustShieldStateCode,
 };
-use matrix_sdk_crypto::CollectStrategy;
 use napi::bindgen_prelude::BigInt;
 use napi_derive::*;
 
@@ -63,14 +62,9 @@ pub struct EncryptionSettings {
     /// created.
     pub history_visibility: events::HistoryVisibility,
 
-    /// Should untrusted devices receive the room key, or should they be
-    /// excluded from the conversation.
-    pub only_allow_trusted_devices: bool,
-
-    /// Should keys be shared with a verified user with an unverified device
-    /// or when a verified user has replaced their identity. Otherwise
-    /// keys are shared with unsigned devices as normal.
-    pub error_on_verified_user_problem: bool,
+    /// The strategy used to distribute the room keys to participant.
+    /// Default will send to all devices.
+    pub sharing_strategy: CollectStrategy,
 }
 
 impl Default for EncryptionSettings {
@@ -90,8 +84,7 @@ impl Default for EncryptionSettings {
                 n.into()
             },
             history_visibility: default.history_visibility.into(),
-            only_allow_trusted_devices: false,
-            error_on_verified_user_problem: false,
+            sharing_strategy: CollectStrategy::AllDevices,
         }
     }
 }
@@ -112,10 +105,89 @@ impl From<&EncryptionSettings> for matrix_sdk_crypto::olm::EncryptionSettings {
             rotation_period: Duration::from_micros(value.rotation_period.get_u64().1),
             rotation_period_msgs: value.rotation_period_messages.get_u64().1,
             history_visibility: value.history_visibility.into(),
-            sharing_strategy: CollectStrategy::DeviceBasedStrategy {
-                only_allow_trusted_devices: value.only_allow_trusted_devices,
-                error_on_verified_user_problem: value.error_on_verified_user_problem,
-            },
+            sharing_strategy: value.sharing_strategy.into(),
+        }
+    }
+}
+
+/// Strategy to collect the devices that should receive room keys for the
+/// current discussion.
+#[napi]
+pub enum CollectStrategy {
+    /// Share with all (unblacklisted) devices.
+    ///
+    /// Not recommended, per the guidance of [MSC4153].
+    ///
+    /// (Used by Element X and Element Web in the legacy, non-"exclude insecure
+    /// devices" mode.)
+    ///
+    /// [MSC4153]: https://github.com/matrix-org/matrix-doc/pull/4153
+    AllDevices,
+
+    /// Share with all devices, except errors for *verified* users cause sharing
+    /// to fail with an error.
+    ///
+    /// In this strategy, if a verified user has an unsigned device,
+    /// key sharing will fail with a
+    /// [`SessionRecipientCollectionError::VerifiedUserHasUnsignedDevice`].
+    /// If a verified user has replaced their identity, key
+    /// sharing will fail with a
+    /// [`SessionRecipientCollectionError::VerifiedUserChangedIdentity`].
+    ///
+    /// Otherwise, keys are shared with unsigned devices as normal.
+    ///
+    /// Once the problematic devices are blacklisted or whitelisted the
+    /// caller can retry to share a second time.
+    ///
+    /// Not recommended, per the guidance of [MSC4153].
+    ///
+    /// [MSC4153]: https://github.com/matrix-org/matrix-doc/pull/4153
+    ErrorOnVerifiedUserProblem,
+
+    /// Share based on identity. Only distribute to devices signed by their
+    /// owner. If a user has no published identity he will not receive
+    /// any room keys.
+    ///
+    /// This is the recommended strategy: it is compliant with the guidance of
+    /// [MSC4153].
+    ///
+    /// (Used by Element Web and Element X in the "exclude insecure devices"
+    /// mode.)
+    ///
+    /// [MSC4153]: https://github.com/matrix-org/matrix-doc/pull/4153
+    IdentityBasedStrategy,
+
+    /// Only share keys with devices that we "trust". A device is trusted if any
+    /// of the following is true:
+    ///     - It was manually marked as trusted.
+    ///     - It was marked as verified via interactive verification.
+    ///     - It is signed by its owner identity, and this identity has been
+    ///       trusted via interactive verification.
+    ///     - It is the current own device of the user.
+    ///
+    /// This strategy is compliant with [MSC4153], but is probably too strict
+    /// for normal use.
+    ///
+    /// (Used by Element Web when "only send messages to verified users" is
+    /// enabled.)
+    ///
+    /// [MSC4153]: https://github.com/matrix-org/matrix-doc/pull/4153
+    OnlyTrustedDevices,
+}
+
+impl Into<matrix_sdk_crypto::CollectStrategy> for CollectStrategy {
+    fn into(self) -> matrix_sdk_crypto::CollectStrategy {
+        match self {
+            CollectStrategy::AllDevices => matrix_sdk_crypto::CollectStrategy::AllDevices,
+            CollectStrategy::ErrorOnVerifiedUserProblem => {
+                matrix_sdk_crypto::CollectStrategy::ErrorOnVerifiedUserProblem
+            }
+            CollectStrategy::IdentityBasedStrategy => {
+                matrix_sdk_crypto::CollectStrategy::IdentityBasedStrategy
+            }
+            CollectStrategy::OnlyTrustedDevices => {
+                matrix_sdk_crypto::CollectStrategy::OnlyTrustedDevices
+            }
         }
     }
 }
@@ -142,8 +214,9 @@ pub enum ShieldStateCode {
     UnsignedDevice,
     /// The sender hasn't been verified by the Client's user.
     UnverifiedIdentity,
-    /// An unencrypted event in an encrypted room.
-    SentInClear,
+    /// The `sender` field on the event does not match the owner of the device
+    /// that established the Megolm session.
+    MismatchedSender,
     /// The sender was previously verified but changed their identity.
     VerificationViolation,
     None,
@@ -158,8 +231,8 @@ impl From<RustShieldStateCode> for ShieldStateCode {
             RustShieldStateCode::UnknownDevice => ShieldStateCode::UnknownDevice,
             RustShieldStateCode::UnsignedDevice => ShieldStateCode::UnsignedDevice,
             RustShieldStateCode::UnverifiedIdentity => ShieldStateCode::UnverifiedIdentity,
-            RustShieldStateCode::SentInClear => ShieldStateCode::SentInClear,
             RustShieldStateCode::VerificationViolation => ShieldStateCode::VerificationViolation,
+            RustShieldStateCode::MismatchedSender => ShieldStateCode::MismatchedSender,
         }
     }
 }